+++ src/messages/layouts/MessageLayoutElement.hpp
@@ -49,12 +49,12 @@ public:
 
     virtual void addCopyTextToString(QString &str, uint32_t from = 0,
                                      uint32_t to = UINT32_MAX) const = 0;
-    virtual int getSelectionIndexCount() const = 0;
+    virtual size_t getSelectionIndexCount() const = 0;
     virtual void paint(QPainter &painter,
                        const MessageColors &messageColors) = 0;
     virtual void paintAnimated(QPainter &painter, int yOffset) = 0;
     virtual int getMouseOverIndex(const QPoint &abs) const = 0;
-    virtual int getXFromIndex(int index) = 0;
+    virtual int getXFromIndex(size_t index) = 0;
 
     const Link &getLink() const;
     const QString &getText() const;
@@ -84,11 +84,11 @@ public:
 protected:
     void addCopyTextToString(QString &str, uint32_t from = 0,
                              uint32_t to = UINT32_MAX) const override;
-    int getSelectionIndexCount() const override;
+    size_t getSelectionIndexCount() const override;
     void paint(QPainter &painter, const MessageColors &messageColors) override;
     void paintAnimated(QPainter &painter, int yOffset) override;
     int getMouseOverIndex(const QPoint &abs) const override;
-    int getXFromIndex(int index) override;
+    int getXFromIndex(size_t index) override;
 
     ImagePtr image_;
 };
@@ -103,11 +103,11 @@ public:
 protected:
     void addCopyTextToString(QString &str, uint32_t from = 0,
                              uint32_t to = UINT32_MAX) const override;
-    int getSelectionIndexCount() const override;
+    size_t getSelectionIndexCount() const override;
     void paint(QPainter &painter, const MessageColors &messageColors) override;
     void paintAnimated(QPainter &painter, int yOffset) override;
     int getMouseOverIndex(const QPoint &abs) const override;
-    int getXFromIndex(int index) override;
+    int getXFromIndex(size_t index) override;
 
     std::vector<ImagePtr> images_;
     std::vector<QSize> sizes_;
@@ -156,11 +156,11 @@ public:
 protected:
     void addCopyTextToString(QString &str, uint32_t from = 0,
                              uint32_t to = UINT32_MAX) const override;
-    int getSelectionIndexCount() const override;
+    size_t getSelectionIndexCount() const override;
     void paint(QPainter &painter, const MessageColors &messageColors) override;
     void paintAnimated(QPainter &painter, int yOffset) override;
     int getMouseOverIndex(const QPoint &abs) const override;
-    int getXFromIndex(int index) override;
+    int getXFromIndex(size_t index) override;
 
     QColor color_;
     FontStyle style_;
@@ -180,11 +180,11 @@ public:
 protected:
     void addCopyTextToString(QString &str, uint32_t from = 0,
                              uint32_t to = UINT32_MAX) const override;
-    int getSelectionIndexCount() const override;
+    size_t getSelectionIndexCount() const override;
     void paint(QPainter &painter, const MessageColors &messageColors) override;
     void paintAnimated(QPainter &painter, int yOffset) override;
     int getMouseOverIndex(const QPoint &abs) const override;
-    int getXFromIndex(int index) override;
+    int getXFromIndex(size_t index) override;
 
 private:
     float scale;
@@ -202,10 +202,10 @@ protected:
     void paint(QPainter &painter, const MessageColors &messageColors) override;
     void paintAnimated(QPainter &painter, int yOffset) override;
     int getMouseOverIndex(const QPoint &abs) const override;
-    int getXFromIndex(int index) override;
+    int getXFromIndex(size_t index) override;
     void addCopyTextToString(QString &str, uint32_t from = 0,
                              uint32_t to = UINT32_MAX) const override;
-    int getSelectionIndexCount() const override;
+    size_t getSelectionIndexCount() const override;
 
 private:
     const QPen pen_;
